+++ src/main.rs
@@ -2,126 +2,118 @@
 mod configuration;
 mod controllers;
 mod skins;
-use controllers::controller::{Controller, Pressed};
+use controllers::controller::Controller;
 
-use ggez::{
-    conf, event,
-    graphics::{self, DrawParam},
-    Context, ContextBuilder, GameResult,
-};
+use sdl2::event::Event;
+use sdl2::image::{InitFlag, LoadTexture};
+use sdl2::keyboard::Keycode;
+use std::collections::HashMap;
+use std::error::Error;
 use rusb2snes::SyncClient;
 use skins::skin::Skin;
-use std::error::Error;
 
 use configuration::config::AppConfig;
 
-const APP_NAME: &str = "Snes Input Display";
-
-struct InputViewer {
-    config: AppConfig,
-    controller: Controller,
-    skin: Skin,
-    client: SyncClient,
-    events: Vec<Pressed>,
-}
-
-impl InputViewer {
-    fn new(ctx: &mut Context, config: AppConfig) -> Result<Self, Box<dyn Error>> {
-        let controller = Controller::new(&config.controller.input_config_path);
-
-        let skin = Skin::new(
-            &config.skin.skins_path,
-            &config.skin.skin_name,
-            &config.skin.skin_theme,
-            ctx,
-        )?;
-
-        /* Connect to USB2SNES Server */
-        let mut client = SyncClient::connect()?;
-
-        client.set_name(String::from(APP_NAME))?;
-
-        let devices = client.list_device()?;
-
-        client.attach(&devices[0])?;
-        let info = client.info()?;
-        println!("Attached to {} - {}", info.dev_type, info.version);
-
-        let window_height = match config.skin.piano_roll {
-            true => skin.background.height * 2.0,
-            false => skin.background.height,
-        };
-
-        // Set the window size
-        ctx.gfx.set_mode(conf::WindowMode {
-            width: skin.background.image.width() as f32,
-            height: window_height,
-            resizable: true,
-            ..Default::default()
-        })?;
-
-        Ok(Self {
-            config,
-            controller,
-            skin,
-            client,
-            events: Vec::new(),
-        })
-    }
-
-    // fn piano_roll(self) {
-    //     se
-    // }
-}
-
-impl event::EventHandler for InputViewer {
-    fn update(&mut self, _ctx: &mut Context) -> GameResult {
-        // Update code here...
-        // const DESIRED_FPS: u32 = 60;
-        self.events = self.controller.pushed(&mut self.client).unwrap();
-        Ok(())
-    }
-
-    fn draw(&mut self, ctx: &mut Context) -> GameResult {
-        let mut canvas = graphics::Canvas::from_frame(ctx, None);
-
-        // Draw background
-        canvas.draw(&self.skin.background.image, DrawParam::new());
-
-        if self.config.skin.piano_roll {
-            self.skin.piano_roll.update(ctx.gfx.size(), &self.events);
-        }
-
-        // Draw inputs
-        for event in self.events.iter() {
-            let button_image = &self.skin.buttons[event].image;
-            canvas.draw(
-                button_image,
-                DrawParam::default().dest(self.skin.buttons[event].rect.point()),
-            );
-            if self.config.skin.piano_roll {
-                for rollrects in self.skin.piano_roll.x_positions.values() {
-                    for rect in rollrects.positions.iter() {
-                        canvas.draw(button_image, DrawParam::new().dest(rect.point()))
-                    }
-                }
-            }
-        }
-
-        canvas.finish(ctx)
-    }
-}
-
-fn main() -> Result<GameResult, Box<dyn Error>> {
+fn main() -> Result<(), Box<dyn Error>> {
     /* Setup Configs */
     let app_config = AppConfig::new()?;
+    
+    let controller = Controller::new(&app_config.controller.input_config_path);
 
-    let (mut ctx, event_loop) = ContextBuilder::new(APP_NAME, "ChTBoner")
-        .add_resource_path(&app_config.skin.skins_path)
-        .window_setup(conf::WindowSetup::default().title(APP_NAME))
+    let skin = Skin::new(&app_config.skin.skins_path, app_config.skin.skin_name);
+
+    /* Connect to USB2SNES Server */
+    let mut usb2snes = SyncClient::connect()?;
+
+
+    usb2snes.set_name(String::from("Snes Input Viewer"))?;
+
+    let devices = usb2snes.list_device()?;
+
+    usb2snes.attach(&devices[0])?;
+    let info = usb2snes.info()?;
+    println!("Attached to {} - {}", info.dev_type, info.version);
+
+    /*
+    Set SDL2 context
+    */
+
+    let sdl_context = sdl2::init()?;
+    let video_subsystem = sdl_context.video()?;
+    let _image_context = sdl2::image::init(InitFlag::PNG | InitFlag::JPG)?;
+    let window = video_subsystem
+        .window(
+            "SNES Input Display",
+            skin.backgrounds[&app_config.skin.skin_theme].width,
+            skin.backgrounds[&app_config.skin.skin_theme].height,
+        )
+        .position_centered()
         .build()
-        .expect("aieee, could not create ggez context!");
+        .map_err(|e| e.to_string())?;
 
-    let input_viewer = InputViewer::new(&mut ctx, app_config)?;
-    event::run(ctx, event_loop, input_viewer);
+    let mut canvas = window
+        .into_canvas()
+        .software()
+        .build()
+        .map_err(|e| e.to_string())?;
+
+    let texture_creator = canvas.texture_creator();
+    // skin.load_textures(&canvas);
+    let background_texture =
+        texture_creator.load_texture(&skin.backgrounds[&app_config.skin.skin_theme].image)?;
+
+    let mut button_textures = HashMap::new();
+
+    button_textures.insert("a", texture_creator.load_texture(&skin.buttons["a"].image)?);
+    button_textures.insert("b", texture_creator.load_texture(&skin.buttons["b"].image)?);
+    button_textures.insert("x", texture_creator.load_texture(&skin.buttons["x"].image)?);
+    button_textures.insert("y", texture_creator.load_texture(&skin.buttons["y"].image)?);
+    button_textures.insert(
+        "select",
+        texture_creator.load_texture(&skin.buttons["select"].image)?,
+    );
+    button_textures.insert(
+        "start",
+        texture_creator.load_texture(&skin.buttons["start"].image)?,
+    );
+    button_textures.insert("r", texture_creator.load_texture(&skin.buttons["r"].image)?);
+    button_textures.insert("l", texture_creator.load_texture(&skin.buttons["l"].image)?);
+    button_textures.insert(
+        "up",
+        texture_creator.load_texture(&skin.buttons["up"].image)?,
+    );
+    button_textures.insert(
+        "down",
+        texture_creator.load_texture(&skin.buttons["down"].image)?,
+    );
+    button_textures.insert(
+        "left",
+        texture_creator.load_texture(&skin.buttons["left"].image)?,
+    );
+    button_textures.insert(
+        "right",
+        texture_creator.load_texture(&skin.buttons["right"].image)?,
+    );
+
+    'mainloop: loop {
+        let events = controller.pushed(&mut usb2snes)?;
+
+        canvas.copy(&background_texture, None, None)?;
+        for event in events {
+            canvas.copy(&button_textures[event], None, skin.buttons[event].rect)?;
+        }
+        canvas.present();
+
+        for event in sdl_context.event_pump()?.poll_iter() {
+            match event {
+                Event::Quit { .. }
+                | Event::KeyDown {
+                    keycode: Option::Some(Keycode::Escape),
+                    ..
+                } => break 'mainloop,
+                _ => {}
+            }
+        }
+    }
+    Ok(())
 }
